modified to show the score on the game over screen too

+++ al1-core/src/main/java/com/gemserk/games/ludumdare/al1/PlayGameState.java
@@ -127,7 +127,7 @@ public class PlayGameState extends GameStateImpl {
 			public void onEvent(Event event) {
 				new TransitionBuilder(game, game.gameOverScreen) //
 						.disposeCurrent() //
-						.parameter("score", score)
+						.parameter("score", (long) score)
 						.start();
 			}
 		});
